+++ test/studio.test.ts
@@ -1310,6 +1310,7 @@ describe('Studio API: home page', () => {
       expect(catalog[page]).not.toContain('main-blog')
       expect(catalog[page]).not.toContain('main-article')
       expect(catalog[page]).not.toContain('main-404')
+      expect(catalog[page]).not.toContain('main-list-collections')
       expect(catalog[page]).not.toContain('predictive-search')
       expect(catalog[page]).not.toContain('quick-add')
     }